old/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
openai-whisper>=20231117
transformers>=4.35.0
torch>=2.1.0
pyannote.audio>=3.1.0
python-ffmpeg>=1.0.3
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
bcrypt>=4.1.1
httpx>=0.25.0
python-dotenv>=1.0.0
pydantic-settings>=2.1.0
huggingface-hub>=0.19.0
aiofiles>=23.2.1
